% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HollandPressureProfilePi}
\alias{HollandPressureProfilePi}
\title{Holland Pressure Profile Time Series}
\usage{
HollandPressureProfilePi(rMax, dP, cP, beta, R)
}
\arguments{
\item{rMax}{radius of maximum winds in km}

\item{dP}{pressure differential, environmental less TC central pressure in hPa}

\item{cP}{TC central pressure in hPa}

\item{beta}{exponential term for Holland vortex}

\item{R}{vector of distances from grid points to TC centre in km}
}
\value{
vector of pressures.
//@example HollandPressureProfilePi(20,20,980,1.2,50)
}
\description{
Pressure profile time series at a grid point.
}
