% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HollandWindProfilePi}
\alias{HollandWindProfilePi}
\title{Holland Wind Profile Time Series}
\usage{
HollandWindProfilePi(f, vMax, rMax, dP, rho, beta, R)
}
\arguments{
\item{f}{single coriolis parameter at the centre of TC in hz}

\item{vMax}{maximum wind velocity calculation in m/s}

\item{rMax}{radius of maximum winds in km}

\item{dP}{pressure differential, environmental less TC central pressure in hPa}

\item{rho}{density of air in Kg/m3}

\item{beta}{exponential term for Holland vortex}

\item{R}{vector of distances from grid points to TC centre in km}
}
\value{
array with two columns for velocity and then vorticity.
//@example HollandWindProfilePi(-1e-4,20,20,10,1.15,1.2,50)
}
\description{
wind profile time series at a grid point
}
