% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindHazaRds.R
\name{inlandWindDecay}
\alias{inlandWindDecay}
\title{Reduce Winds Overland}
\usage{
inlandWindDecay(d, a = c(0.66, 1, 0.4))
}
\arguments{
\item{d}{inland distance in km}

\item{a}{three parameter of decay model a1,a2,a3}
}
\value{
a reduction factor Km
}
\description{
Reduce Winds Overland
}
\examples{
inlandWindDecay(10)
}
