% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{JelesnianskiWindProfilePi}
\alias{JelesnianskiWindProfilePi}
\title{Jelesnianski Wind Profile Time Series}
\usage{
JelesnianskiWindProfilePi(f, vMax, rMax, R)
}
\arguments{
\item{f}{single coriolis parameter at the centre of TC in hz}

\item{vMax}{maximum wind velocity calculation in m/s}

\item{rMax}{radius of maximum winds in km}

\item{R}{vector of distances from grid points to TC centre in km}
}
\value{
array with two columns for velocity and then vorticity.
//@example JelesnianskiWindProfilePi(-1e-4,20,20,50)
}
\description{
wind profile time series at a grid point
}
