% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rdist}
\alias{Rdist}
\title{TC Distance and Direction From Output Grid Points}
\usage{
Rdist(Gridlon, Gridlat, TClon, TClat)
}
\arguments{
\item{Gridlon}{vector of Grid point longitudes}

\item{Gridlat}{vector of Grid point latitudes}

\item{TClon}{single TC longitude}

\item{TClat}{single TC latitude}
}
\value{
two columns for distance in km and cartesian direction in degrees, counter clockwise from the x axis.
//@example Rdist(c(144,145),c(-11,-12),142,-14)
}
\description{
Grid points distance and direction to TC.
}
