% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdistPi}
\alias{RdistPi}
\title{TC Track Distance and Direction From Output Grid Point}
\usage{
RdistPi(Gridlon, Gridlat, TClon, TClat)
}
\arguments{
\item{Gridlon}{single Grid point longitude}

\item{Gridlat}{single Grid point latitude}

\item{TClon}{vector of TC longitudes}

\item{TClat}{vector of TC latitudes}
}
\value{
two columns for distance in km and cartesian direction in degrees, counterclockwise from the x axis.
//@example RdistPi(142,-14,c(144,145),c(-11,-12))
}
\description{
Grid point time series of TC distance and direction.
}
