% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindHazaRds.R
\name{readncs}
\alias{readncs}
\title{Title Read All netcdf Variables Into a list()}
\usage{
readncs(infile)
}
\arguments{
\item{infile}{Name of the existing netCDF file to be opened.}
}
\value{
a list of lists with netCDF variables read by the ncdf4 package.
}
\description{
Title Read All netcdf Variables Into a list()
}
